Ensure scripts are all read at the same revision
Get the HEAD SHA of the selected branch, then use that SHA instead
of the branch name when downloading. This ensures all scripts are
downloaded at the same revision. Without this, it was possible for
the additional scripts to be of different versions if the repo
happened to be updated while setup was in progress.