Save hourly snapshots of disk image
When archiveloop starts, and every hour thereafter, take a snapshot
of the disk image. This allows keeping more than the last hour of
recordings, without the overhead of actually making a copy of
every file, which would take took long.
There is no way to access or archive the snapshots, yet.